FDG-5406 add bar animation
