+++ package.json
@@ -53,6 +53,7 @@
     "@nivo/line": "^0.79.1",
     "@nivo/tooltip": "^0.79.0",
     "@reach/router": "^1.3.4",
+    "@react-spring/web": "^9.7.1",
     "@types/react": "^17.0.14",
     "@types/react-dom": "^17.0.9",
     "awesome-debounce-promise": "^2.1.0",
