+++ src/layouts/explainer/sections/national-debt/breaking-down-the-debt/intragovernmental-holdings-chart/intragovernmental-holdings-chart.module.scss
@@ -1,0 +1,26 @@
+@import '../../../../../../variables.module.scss';
+
+
+.fillBar {
+
+}
+
+:global(.debtBar){
+  //transform: scaleX(0);
+  //transition: all 6s;
+  transition: transform 3s 3s ease-in-out;
+  transform: scaleY(0);
+  //height: 0;
+  //  transform-origin: 5%;
+  &:global(.fillRight) {
+    //height: 100%;
+    transform: scaleY(1);
+    animation-direction: reverse;
+  }
+}
+
+.markerText {
+  font-size: $font-size-36;
+  font-weight: $bold-weight;
+  color: #666666;
+}
