+++ src/layouts/explainer/sections/national-debt/breaking-down-the-debt/intragovernmental-holdings-chart/intragovernmental-holdings-chart.module.scss.d.ts
@@ -1,0 +1,66 @@
+// This file is automatically generated. Do not modify this file manually -- YOUR CHANGES WILL BE ERASED!
+export const breakpointSm: string;
+export const breakpointMd: string;
+export const breakpointLg: string;
+export const breakpointXl: string;
+export const siteContentWidth: string;
+export const fontTitle: string;
+export const fontBodyCopy: string;
+export const fontSize_10: string;
+export const fontSize_12: string;
+export const fontSize_13: string;
+export const fontSize_14: string;
+export const fontSize_15: string;
+export const fontSize_16: string;
+export const fontSize_18: string;
+export const fontSize_20: string;
+export const fontSize_22: string;
+export const fontSize_24: string;
+export const fontSize_28: string;
+export const fontSize_32: string;
+export const fontSize_36: string;
+export const fontSize_40: string;
+export const fontSize_48: string;
+export const fontSize_56: string;
+export const apiHeader: string;
+export const placeholder: string;
+export const darkSlateBlue: string;
+export const slateBlue: string;
+export const linkColor: string;
+export const primary: string;
+export const primaryDarker: string;
+export const primaryInactive: string;
+export const primaryGlass: string;
+export const primaryLightBackground: string;
+export const darkPink: string;
+export const dropdownActiveBackground: string;
+export const bodyBackground: string;
+export const contentSectionBackground: string;
+export const borderColor: string;
+export const errorBackground: string;
+export const errorTextColor: string;
+export const errorLinkColor: string;
+export const lightWeight: string;
+export const semiBoldWeight: string;
+export const boldWeight: string;
+export const primaryPale: string;
+export const disabledButtonBg: string;
+export const helpTextColor: string;
+export const animationDuration: string;
+export const darkGreen: string;
+export const darkGreyBorder: string;
+export const infoBoxBackground: string;
+export const infoBoxBorder: string;
+export const basicWarningBackground: string;
+export const basicWarningBorder: string;
+export const focus: string;
+export const cardBody: string;
+export const mainWidth: string;
+export const ddBorderColor: string;
+export const altButtonActiveBg: string;
+export const infoIcon: string;
+export const paleBg: string;
+export const debtExplainerPrimary: string;
+export const debtExplainerSecondary: string;
+export const debtExplainerLightSecondary: string;
+export const markerText: string;
