+++ src/layouts/explainer/sections/national-debt/national-debt.module.scss
@@ -574,10 +574,7 @@
 
     .footerContainer {
       padding-top: 15px;
-
-      p {
-        font-size: $font-size-12;
-      }
+      font-size: $font-size-12;
     }
   }
